+++ kte/src/commonMain/kotlin/com/wakaztahir/kte/parser/stream/DestinationStream.kt
@@ -1,7 +1,7 @@
 package com.wakaztahir.kte.parser.stream
 
 import com.wakaztahir.kte.model.BooleanValue
-import com.wakaztahir.kte.model.FloatValue
+import com.wakaztahir.kte.model.DoubleValue
 import com.wakaztahir.kte.model.IntValue
 import com.wakaztahir.kte.model.StringValue
 
@@ -11,7 +11,7 @@ abstract class DestinationStream {
     abstract fun write(str: String)
 
     open fun write(value: IntValue) = write(value.value.toString())
-    open fun write(value: FloatValue) = write(value.value.toString() + 'f')
+    open fun write(value: DoubleValue) = write(value.value.toString() + 'f')
     open fun write(value: BooleanValue) = write(if (value.value) "true" else "false")
     open fun write(value: StringValue) = write(value.value)
